+++ packages/preact/src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useForm';
 export * from './useFormStore';
 export * from './useLifecycle';
+export * from './useLiveSignal';