+++ packages/react/src/components/FieldArray.tsx
@@ -1,6 +1,6 @@
 import type { ReadonlySignal } from '@preact/signals-react';
 import { type ReactNode, useMemo } from 'react';
-import { useLifecycle } from '../hooks';
+import { useLifecycle, useLiveSignal } from '../hooks';
 import type {
   FieldArrayPath,
   FieldValues,
@@ -68,19 +68,28 @@ export function FieldArray<
   // Use lifecycle of field array
   useLifecycle({ ...props, store: fieldArray });
 
+  // Create readonly live signals
+  // https://github.com/preactjs/signals/issues/361
+  const items = useLiveSignal(fieldArray.items);
+  const error = useLiveSignal(fieldArray.error);
+  const active = useLiveSignal(fieldArray.active);
+  const touched = useLiveSignal(fieldArray.touched);
+  const dirty = useLiveSignal(fieldArray.dirty);
+
   return (
     <>
       {children(
         useMemo(
           () => ({
             name,
-            items: fieldArray.items,
-            error: fieldArray.error,
-            active: fieldArray.active,
-            touched: fieldArray.touched,
-            dirty: fieldArray.dirty,
+            items,
+            error,
+            active,
+            touched,
+            dirty,
           }),
-          [fieldArray, name]
+          // eslint-disable-next-line react-hooks/exhaustive-deps
+          [name]
         )
       )}
     </>