+++ packages/react/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Fix bug when using dynamic field arrays
+
 ## v0.1.0 (May 15, 2023)
 
 - Initial release